# packages/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Eigen Engine API",
    "version": "1.0.0",
    "description": "The server-authoritative game engine API. Client routes (`/api/engine/*`) require a Firebase ID token; the external-bot webhook (`/api/bot/action`) is HMAC-authenticated (§7)."
  },
  "security": [
    {
      "firebase": []
    }
  ],
  "components": {
    "securitySchemes": {
      "firebase": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A Firebase ID token, sent as `Authorization: Bearer <token>` (or `?token=` on WebSocket upgrades)."
      },
      "botHmac": {
        "type": "apiKey",
        "in": "header",
        "name": "Eigen-Signature",
        "description": "An external bot's HMAC signature over the exact request body, bound to the `action` domain (§7). Scheme `v1,<base64>`; the per-bot key is `HMAC(BOT_SIGNING_SECRET, bot_id)`. The engine signs wakes with the same header in the other direction."
      }
    },
    "schemas": {
...
      "BotActionAccepted": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "ok"
        ]
      },
      "BotAction": {
        "type": "object",
        "properties": {
          "bot_id": {
            "type": "string",
            "minLength": 1
          },
          "game_id": {
            "type": "string",
            "minLength": 1
          },
          "player_index": {
            "type": "integer",
            "minimum": 0
          },
          "version": {
            "type": "integer",
            "minimum": 0
          },
          "data": {}
        },
        "required": [
          "bot_id",
          "game_id",
          "player_index",
          "version"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/bot/action": {
      "post": {
        "operationId": "botAction",
        "security": [
          {
            "botHmac": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BotAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The move was accepted and applied",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotActionAccepted"
                }
              }
            }
          },
          "400": {
            "description": "Malformed body",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Invalid signature",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
...
}